#262 Reduce memory needed for time related objects
Simplify RelativeTime and SimpleTime to reduce memory.

+++ mobitopp/src/main/java/edu/kit/ifv/mobitopp/populationsynthesis/serialiser/DefaultActivityFormat.java
@@ -83,7 +83,7 @@ public class DefaultActivityFormat implements SerialiserFormat<PersonPatternActi
 	}
 
 	private Time startTimeOf(List<String> data) {
-		long secondsFromStart = Long.parseLong(data.get(startTimeIndex));
+		int secondsFromStart = Integer.parseInt(data.get(startTimeIndex));
 		return SimpleTime.ofMinutes(secondsFromStart);
 	}
